requirements.txt
absl-py==1.2.0
anyconfig==0.13.0
anyio==3.6.1
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
boto3==1.24.88
botocore==1.27.88
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
cloudpickle==2.2.0
colorama==0.4.5
connexion==2.14.1
contextlib2==21.6.0
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
decorator==5.1.1
executing==1.1.1
fastapi==0.85.0
Flask==2.2.2
Flask-OpenTracing==1.1.0
flatbuffers==22.9.24
fonttools==4.37.4
future==0.18.2
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h11==0.14.0
h5py==3.7.0
idna==3.4
imageio==2.22.1
importlib-metadata==5.0.0
inflection==0.5.1
ipython==8.5.0
itsdangerous==2.1.2
jaeger-client==4.8.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lime==0.2.0.1
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
networkx==2.8.7
numba==0.56.3
numpy==1.23.3
oauthlib==3.2.1
openapi==1.1.0
opencv-python==4.6.0.66
opentracing==2.4.0
opentracing-instrumentation==3.3.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prance==0.21.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
pure-eval==0.2.2
py-ms==2.8.0
py-ms-consulate==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-multipart==0.0.5
pytz==2022.5
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
semver==2.13.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
sniffio==1.3.0
stack-data==0.5.1
starlette==0.20.4
swagger-ui-bundle==0.0.9
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tf==1.0.0
tf-explain==0.3.1
threadloop==1.0.2
threadpoolctl==3.1.0
thrift==0.16.0
tifffile==2022.10.10
tornado==5.1.1
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
